+++ src/core/cmd_msg_bar.h
@@ -16,28 +16,28 @@ class Choices {
 public:
   Choices(ChoicesFilter cf): filter(cf), optLoc(-1) {}
   virtual ~Choices() {}
-  virtual const std::string& at(int idx) const = 0;
-  virtual int size() const = 0;
+  virtual const std::string& at(size_t idx) const = 0;
+  virtual size_t size() const = 0;
   virtual bool updateChoices(const std::string& str) { return false; }
-  virtual std::string getFinalStr(int idx, const std::string& str) const = 0;
+  virtual std::string getFinalStr(size_t idx, const std::string& str) const = 0;
   virtual void updateMainBuffer(CmdMsgBar& cmBar) {}
   bool match(const std::string& line, const std::string& str) const;
-  bool match(int idx, const std::string& str) const;
-  void setIdx(int idx) { optLoc = idx; }
-  int getIdx() const { return optLoc; }
+  bool match(size_t idx, const std::string& str) const;
+  void setIdx(size_t idx) { optLoc = idx; }
+  size_t getIdx() const { return optLoc; }
 
 private:
   ChoicesFilter filter;
-  int optLoc;
+  size_t optLoc;
 };
 
 
 class StringChoices: public Choices {
 public:
   StringChoices(const Strings& arr, ChoicesFilter cf=strFind);
-  const std::string& at(int idx) const { return options[idx]; }
-  std::string getFinalStr(int idx, const std::string& str) const;
-  int size() const { return (int)options.size(); }
+  const std::string& at(size_t idx) const { return options[idx]; }
+  std::string getFinalStr(size_t idx, const std::string& str) const;
+  size_t size() const { return options.size(); }
 
 protected:
   Strings options;
@@ -50,14 +50,14 @@ public:
   void insert(char c) override;
   void insert(const std::string& str) override;
   void draw(Editor& ed, const Window& win) override;
-  int drawLine(int y, const std::string& line, Editor& ed, int lineNum,
-               const Window& win) override;
-  void load(const std::string& file, int line=0) override {}
+  size_t drawLine(size_t y, const std::string& line, Editor& ed, size_t lineNum,
+                  const Window& win) override;
+  void load(const std::string& file, size_t line = 0) override {}
   bool save(const std::string& fName="") override { return false; }
   void clear() override;
-  void setMinLoc(int loc) { minLoc = loc; }
-  int getMinStartLoc() const override { return minLoc; }
-  int totalLinesNeeded(const Point& dim) const override;
+  void setMinLoc(size_t loc) { minLoc = loc; }
+  size_t getMinStartLoc() const override { return minLoc; }
+  size_t totalLinesNeeded(const Point& dim) const override;
   void setChoices(Choices* ch) { choices = ch; }
   void clearChoices();
   bool usingChoices() const { return choices != nullptr; }
@@ -66,19 +66,19 @@ public:
   void down();
   void up();
   void updateChoices();
-  int getOptLoc() const { return optLoc; }
+  size_t getOptLoc() const { return optLoc; }
   void lineUp(const Point& dim) override;
   void lineDown() override;
 
 private:
   /** useful during prompts, so as to not cross into the message itself! */
-  int minLoc;
+  size_t minLoc;
   /** external choices vector (usually options) that need to be rendered */
   Choices* choices;
   /** currently selected option */
-  int optLoc;
+  size_t optLoc;
 
-  int linesNeeded(const std::string& str, int wid) const;
+  size_t linesNeeded(const std::string& str, size_t wid) const;
 };
 
 } // end namespace teditor
