+++ src/core/isearch.h
@@ -20,20 +20,20 @@ class ISearch: public Choices {
 public:
   ISearch(Window& w, bool _noCase);
 
-  const std::string& at(int idx) const;
-  int size() const { return ml.length(); }
+  const std::string& at(size_t idx) const;
+  size_t size() const { return ml.length(); }
   bool updateChoices(const std::string& str);
-  std::string getFinalStr(int idx, const std::string& str) const;
+  std::string getFinalStr(size_t idx, const std::string& str) const;
   void updateMainBuffer(CmdMsgBar& cmBar);
 
   /** reset the search state */
   void reset();
 
   /** check whether there are matches at a given line num */
-  bool emptyAt(int i) const { return matches.find(i) == matches.end(); }
+  bool emptyAt(size_t i) const { return matches.find(i) == matches.end(); }
 
   /** fetch matches for a given line num */
-  const std::vector<int>& matchesAt(int i) const;
+  const std::vector<size_t>& matchesAt(size_t i) const;
 
 private:
   /** current window which has this buffer */
@@ -43,13 +43,13 @@ private:
   /** current search string */
   std::string curr;
   /** line number v/s list of matches */
-  std::unordered_map<int, std::vector<int>> matches;
+  std::unordered_map<size_t, std::vector<size_t>> matches;
   /** case insensitive search? */
   bool noCase;
 
   void searchBuffer();
-  void searchLine(const std::string& str, std::vector<int>& res);
-  void iSearchLine(const std::string& str, std::vector<int>& res);
+  void searchLine(const std::string& str, std::vector<size_t>& res);
+  void iSearchLine(const std::string& str, std::vector<size_t>& res);
 };
 
 } // end namespace teditor
