+++ src/core/line.h
@@ -21,37 +21,37 @@ public:
   void append(const std::string& str) { line += str; }
   void prepend(char c) { insert(c, 0); }
   void prepend(const char* c) { insert(c, 0); }
-  void prepend(char c, int count);
-  void insert(char c, int idx);
-  void insert(const char* c, int idx);
-  void insert(const std::string& str, int idx);
+  void prepend(char c, size_t count);
+  void insert(char c, size_t idx);
+  void insert(const char* c, size_t idx);
+  void insert(const std::string& str, size_t idx);
   /** @} */
 
   /** erases from the given index and returns the erased chars */
-  std::string erase(int idx, int len=1);
+  std::string erase(size_t idx, size_t len = 1);
 
   /**
    * @defgroup SplitJoin Functions to split and join the line back
    * @{
    */
-  Line split(int idx);
+  Line split(size_t idx);
   void join(const Line& other);
   /** @} */
 
   /** number of lines needed on the screen to render this line */
-  int numLinesNeeded(int wid) const;
+  size_t numLinesNeeded(size_t wid) const;
 
   /** Check for empty line */
   bool empty() const { return line.empty(); }
 
   /** Number of chars in the line */
-  int length() const { return (int)line.length(); }
+  size_t length() const { return line.length(); }
 
   /** get the string */
   const std::string& get() const { return line; }
 
   /** access idx'th element in the line */
-  char at(int idx) const { return line[idx]; }
+  char at(size_t idx) const { return line[idx]; }
 
   /** clear the line */
   void clear() { line.clear(); }
@@ -60,16 +60,16 @@ public:
    * Same as find_first_not_of function of std::string
    * However, if no match, it will return the length of the line itself
    */
-  int findFirstNotOf(const std::string& str, int pos) const;
+  size_t findFirstNotOf(const std::string& str, size_t pos) const;
 
   /**
    * Same as find_last_not_of function of std::string
    * However, if no match, it will return 0
    */
-  int findLastNotOf(const std::string& str, int pos) const;
+  size_t findLastNotOf(const std::string& str, size_t pos) const;
 
   /** measures the indentation at the beginning of this line */
-  int indentSize() const;
+  size_t indentSize() const;
 
 private:
   /** string to be present on this line */
